# openapi/ward.v0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ward API",
    "version": "0.1.0",
    "description": "Ward is tenant containment for AI SaaS: an OpenAI-compatible egress\nproxy plus a tenant control plane. Route agent/tool calls through\nWard with a tenant ID header; operators constrain, pause, resume,\nand audit one tenant while the rest keep running.\n\nHTTP/OpenAPI is Ward's primary integration contract. This document\ndescribes the implemented v0 surface.\n\nHonesty notes (prototype status):\n- Control auth is a single shared bearer token (prototype), NOT\n  production RBAC. It guards mutating /ward/* routes only when\n  WARD_REQUIRE_CONTROL_TOKEN=true; read endpoints are open.\n- POST /v1/chat/completions is OpenAI-compatible enough for the\n  demo and simple integrations; it is not a full OpenAI spec clone\n  (no tool-call passthrough guarantees in mock mode, no SSE\n  pass-through support; deterministic mock streaming only).\n  dev-only prototype disabled by default; kubernetes is planned.\n- Single-node, prototype/demo-supported operation. Production\n  readiness is not claimed.\n"
  },
  "servers": [
    {
      "url": "http://localhost:4317",
      "description": "Local/self-hosted Ward"
    }
  ],
...
  "paths": {
...
    "/ward/tenants/{tenantId}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "operationId": "getTenant",
        "summary": "Inspect one tenant",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant with current pressure evaluation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "tenant"
                  ],
                  "properties": {
                    "tenant": {
                      "$ref": "#/components/schemas/Tenant"
                    },
                    "pressure": {
                      "type": "object",
                      "properties": {
                        "detected": {
                          "type": "boolean"
                        },
                        "requestsInWindow": {
                          "type": "integer"
                        },
                        "windowMs": {
                          "type": "integer"
                        },
                        "threshold": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tenant not tracked yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/constrain": {
      "post": {
        "tags": [
          "tenants"
        ],
        "operationId": "constrainTenant",
        "summary": "Constrain a tenant (direct; control auth when enabled)",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TransitionResult"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/pause": {
      "post": {
        "tags": [
          "tenants"
        ],
        "operationId": "pauseTenant",
        "summary": "Pause a tenant (direct; control auth when enabled)",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TransitionResult"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/resume": {
      "post": {
        "tags": [
          "tenants"
        ],
        "operationId": "resumeTenant",
        "summary": "Resume a tenant (direct; control auth when enabled)",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TransitionResult"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/mode": {
      "post": {
        "tags": [
          "tenants"
        ],
        "operationId": "setTenantModeOverride",
        "summary": "Set a tenant's per-tenant mode override (control auth when enabled)",
        "description": "Override the global WARD_MODE for a single tenant. Supported\nvalues: \"inherit\" (defer to global WARD_MODE), \"observe\" (force\nobserve mode for this tenant), \"enforce\" (force enforce mode for\nthis tenant). There is no \"off\" or silent bypass value — \"observe\"\nstill emits would-block audit evidence when state would have\nblocked. Override changes are audited with action\n\"mode_override_changed\".\n",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TransitionResult"
          },
          "400": {
            "description": "Invalid mode override value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/approval": {
      "post": {
        "tags": [
          "approvals"
        ],
        "operationId": "requestApproval",
        "summary": "Request a single-use approval token for a containment action",
        "description": "Issues a short-lived (default 120s), single-use token bound to\nthe tenant and action, plus the exact confirmation phrase the\noperator must echo back. Prototype operator boundary.\n",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/ward/tenants/{tenantId}/apply-approved-action": {
      "post": {
        "tags": [
          "approvals"
        ],
        "operationId": "applyApprovedAction",
        "summary": "Consume an approval token by echoing the exact phrase",
        "security": [
          {
            "controlToken": []
          },
          {}
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyApprovedActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TransitionResult"
          },
          "400": {
            "description": "Token not found / already used / expired / tenant mismatch / phrase mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WardError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "controlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Minimal shared-token control auth PROTOTYPE (not production\nRBAC). Required on mutating /ward/* routes only when the API\nruns with WARD_REQUIRE_CONTROL_TOKEN=true; otherwise these\nroutes are open (the default for local demos, loudly flagged\nin /health).\n"
      }
    },
    "parameters": {
...
      "TenantIdPath": {
        "name": "tenantId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Control auth required and token missing/invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WardError"
            }
          }
        }
      },
      "TransitionResult": {
        "description": "Tenant after the transition plus the audit event",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "tenant",
                "audit"
              ],
              "properties": {
                "tenant": {
                  "$ref": "#/components/schemas/Tenant"
                },
                "audit": {
                  "$ref": "#/components/schemas/AuditEvent"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "TenantState": {
        "type": "string",
        "enum": [
          "running",
          "constrained",
          "paused"
        ]
      },
      "WardModeOverride": {
        "description": "Per-tenant mode override. \"inherit\" defers to the global\nWARD_MODE setting; \"observe\" and \"enforce\" pin the tenant\nregardless of the global default. There is no \"off\" or silent\nbypass value.\n",
        "type": "string",
        "enum": [
          "inherit",
          "observe",
          "enforce"
        ]
      },
      "Tenant": {
        "type": "object",
        "required": [
          "tenantId",
          "state",
          "requestCount",
          "detectedPressure",
          "estimatedSpend",
          "activeWorkflowRuns",
          "deploymentMode",
          "modeOverride",
          "updatedAt"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TenantState"
          },
          "requestCount": {
            "type": "integer"
          },
          "recentRequestTimestamps": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "detectedPressure": {
            "type": "boolean"
          },
          "lastPressureReason": {
            "type": "string",
            "nullable": true
          },
          "estimatedSpend": {
            "type": "number",
            "description": "Fixed per-call estimate; NOT a real cost model."
          },
          "activeWorkflowRuns": {
            "type": "integer"
          },
          "deploymentMode": {
            "type": "string",
            "enum": [
              "local",
              "docker",
              "kubernetes",
              "hosted"
            ]
          },
          "modeOverride": {
            "$ref": "#/components/schemas/WardModeOverride"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TenantModeRequest": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/WardModeOverride"
          },
          "actor": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "AuditEvent": {
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "tenantId",
          "action",
          "actor",
          "reason"
        ],
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "tenantId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "previousState": {
            "type": "string"
          },
          "nextState": {
            "type": "string"
          },
          "evidence": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ApprovalRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "constrain",
              "pause",
              "resume"
            ]
          },
          "actor": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "ApprovalResponse": {
        "type": "object",
        "required": [
          "approvalToken",
          "requiredPhrase",
          "expiresAt",
          "action",
          "tenantId"
        ],
        "properties": {
          "approvalToken": {
            "type": "string"
          },
          "requiredPhrase": {
            "type": "string",
            "description": "Must be echoed back byte-for-byte in apply-approved-action."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "action": {
            "type": "string",
            "enum": [
              "constrain",
              "pause",
              "resume"
            ]
          },
          "tenantId": {
            "type": "string"
          }
        }
      },
      "ApplyApprovedActionRequest": {
        "type": "object",
        "required": [
          "approvalToken",
          "phrase"
        ],
        "properties": {
          "approvalToken": {
            "type": "string"
          },
          "phrase": {
            "type": "string"
          }
        }
      },
      "TransitionRequest": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "WardError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. ward_tenant_constrained, ward_control_unauthorized."
          },
          "message": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TenantState"
          }
        }
      },
...
    }
  }
}